Removes unecessary term traversal 't' library.
