+++ package.json
@@ -8,13 +8,13 @@
     "build-prod": "cross-env NODE_ENV=production webpack --config webpack.prod.js --progress --mode production"
   },
   "dependencies": {
-    "apexcharts": "^3.37.1",
+    "apexcharts": "^3.39.0",
     "axios": "^1.1.3",
     "blurhash": "^2.0.5",
     "buefy": "^0.9.23",
     "bulma": "^0.9.4",
     "conditioner-core": "^2.3.3",
-    "countup.js": "^2.5.0",
+    "countup.js": "^2.6.0",
     "css-vars-ponyfill": "^2.4.8",
     "floating-vue": "^1.0.0-beta.19",
     "leaflet": "^1.9.3",
@@ -22,12 +22,11 @@
     "masonry-layout": "^4.2.2",
     "moment": "^2.29.4",
     "node-sass": "^7.0.3",
-    "photoswipe": "^5.3.6",
+    "photoswipe": "^5.3.7",
     "qs": "^6.11.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "swiper": "^8.4.5",
-    "t": "^0.5.1",
     "vue": "^2.6.14",
     "vue-apexcharts": "^1.6.2",
     "vue-blurhash": "^0.1.4",
