+++ src/views/admin/js/store/modules/taxonomy/mutations.js
@@ -1,5 +1,4 @@
 import Vue from 'vue';
-import t from 't';
 
 export const setRepositoryTotalTaxonomies = (state, repositoryTotalTaxonomies) => {
     state.repositoryTotalTaxonomies = repositoryTotalTaxonomies;
@@ -30,11 +29,10 @@ export const deleteTaxonomy = ( state, taxonomy ) => {
 export const setSingleTerm = (state, term) => {
 
     let index = state.terms.findIndex(updatedTerm => updatedTerm.id === term.id);
-    if ( index >= 0){
+    if ( index >= 0 )
         Vue.set( state.terms, index, term );
-    } else {
+    else
         state.terms.push( term );
-    }
 };
 
 export const setTerms = (state, terms) => {
@@ -47,124 +45,6 @@ export const setTerms = (state, terms) => {
             state.terms.push(term);
     }
 
-};
-
-export const clearTerms = (state) => {
-    state.terms = [];
-};
-
-// Hierarchy usage of terms list -----------------
-export const setChildTerms = (state, { terms, parent }) => {
-    
-    if (parent > 0 ) {
-        for (let i = 0; i < state.terms.length; i++) {
-            let parentTerm = t.find(state.terms[i], [], (node, par) => { return node.id == parent; });
-            if (parentTerm != undefined) {
-                if (parentTerm['children'] == undefined)
-                    Vue.set(parentTerm, 'children', []);
-
-                for (let term of terms){
-                    let existingTermIndex = parentTerm['children'].findIndex(aTerm => aTerm.id == term.id);
-                    if (existingTermIndex < 0)
-                        parentTerm['children'].push(term);
-                    else
-                        Vue.set(parentTerm['children'], existingTermIndex, term);
-                }    
-            }
-        }
-    } else {
-        if (state.terms != undefined) {   
-            for (let term of terms) {
-                let existingTermIndex = state.terms.findIndex(aTerm => aTerm.id == term.id);
-                if (existingTermIndex < 0)
-                    state.terms.push(term);
-                else
-                    Vue.set(state.terms, existingTermIndex, term);
-            }
-        } else {    
-            state.terms = terms;
-        }
-    }
-};
-
-export const addChildTerm = (state, { term, parent }) => {
-    
-    if (parent > 0 ) {
-        for (let aTerm of state.terms) {
-            let parentTerm = t.find(aTerm, [], (node, par) => { return node.id == parent; });
-            if (parentTerm != undefined) {
-                if (parentTerm['children'] == undefined) {
-                    Vue.set(parentTerm, 'children', []);
-                }
-                parentTerm['children'].unshift(term);  
-                parentTerm.total_children = parentTerm.children.length;
-            }
-        }
-    } else {
-        if (state.terms != undefined) {
-            let existingTermIndex = state.terms.findIndex(aTerm => aTerm.id == term.id);
-            if (existingTermIndex >= 0)
-                Vue.set(state.terms, existingTermIndex, term);
-            else
-                state.terms.unshift(term);
-        } else {    
-            state.terms = []
-            state.terms.unshift(term);
-        }
-    }
-};
-
-export const updateChildTerm = (state, { term, parent, oldParent }) => {
-
-    if (oldParent == undefined) {
-        if (parent > 0 ) {
-            for (let aTerm of state.terms) {
-                let childTerm = t.find(aTerm, [], (node, par) => { return node.id == term.id; });
-                if (childTerm != undefined) {
-                    childTerm = term;   
-                }
-            }
-        } else {
-            if (state.terms != undefined) {
-                for (let i = 0; i < state.terms.length; i++) {
-                    if (state.terms[i].id == term.id)
-                        Vue.set(state.terms, i, term);
-                }
-            } else {    
-                state.terms = []
-                state.terms.push(term);
-            }
-        }
-    } else {
-        
-        // Removes from old parent
-        deleteChildTerm(term.id, oldParent)
-        // Adds it to new one
-        addChildTerm(term, parent);
-    }
-};
-
-export const deleteChildTerm = ( state, {termId, parent} ) => {
-    
-    if (parent > 0 ) {
-        for (let i = 0; i < state.terms.length; i++) {
-            let parentTerm = t.find(state.terms[i], [], (node, par) => { return node.id == parent; });
-            if (parentTerm != undefined) {
-                let index = parentTerm.children.findIndex(deletedTerm => deletedTerm.id == termId);
-                if (index >= 0) {
-                    parentTerm.children.splice(index, 1);
-                    parentTerm.total_children = parentTerm.children.length;
-                }
-            }
-        }
-    } else {
-        if (state.terms != undefined) {
-            for (let i = 0; i < state.terms.length; i++) {
-                if (state.terms[i].id == termId)
-                    state.terms.splice(i, 1);
-            }
-        }
-    }
 };
 
 export const deleteTerm = ( state, termId ) => {
